requirements.txt
requests==2.28.2
rich==13.5.2
ffmpeg-python==0.2.0
grequests==0.7.0
DownloadKit==1.0.4
chompjs==1.2.4
bs4==0.0.1
beautifulsoup4==4.12.2